Login-Show_companies
In this push, there will be a header for the website, is only a V 0 to make it fonctional with the rest.
Also the login is working and it's possible to see the list of companies.
But the feature of seeing the companies individually is not implemented neither are the offers or the admin part.
Also the migration is working but the seeder have a problem with half of the database