CRUD USER
In this push there are the features for the CRUD of the users except the research that will be done later on.
The Delete is only allowed to the admins and they can't delete themself nor other admins. And the teacher can only create user accounts.
The restriction in the moment of creation and moficiation of the account are still to be reviewed with their respective error messages